+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## dense 1.1.0  not yet released
 
+* Add Dense.path(path) and Dense.gather(collection, path)
 * Use but enhance KeyError and TypeError
 * Differentiate between `*` and `.*`
 * Complete ework around Path#gather